move setup code to subpackage
